Issue #335: move TagCache.h to lib/model/tag folder
