+++ lib/model/FavoritesModel.h
@@ -22,8 +22,8 @@
 #include "FavoritesModelItem.h"
 #include "NoteCache.h"
 #include "SavedSearchCache.h"
-#include "TagCache.h"
 #include "notebook/NotebookCache.h"
+#include "tag/TagCache.h"
 
 #include <quentier/local_storage/LocalStorageManagerAsync.h>
 #include <quentier/types/Account.h>
